far from perfect, but a quick patch to sort of support Styled JSX a little
this change adds delegation to css highlighting for the content of an
interpolated JSX value which is a template string which appears in a
style element that has the jsx attribute. It’s far from perfect —
mainly because the css def isn’t adequate to account for the presence
of ES interpolated sequences (can’t "resume"). In order to do this
right, we’ll need to include our own version of the css def that we can
make happier about arbitrary entry/exit, but I figured something was
better than nothing.